+++ src/Platform/Kernel/Dashboard.php
@@ -11,7 +11,7 @@ class Dashboard
     /**
      * Orchid Version.
      */
-    const VERSION = '2.0.4';
+    const VERSION = '2.0.5';
 
     /**
      * @var